perf: rasterizer optimization
I did a lot of different things that allowed for optimization and do not
remember them all. I know I should've pushed them one at a time, but I
didn't feel like it :( It improved debug speed by around 2x, and my fps are
capped in release mode anyways and I'm too lazy to add stuff on the
screen, so I didn't benchmark it there.

From what I can remember, I:
  - Improved the from_uv method in texture.
  - Removed complicated operations from the big loop in rasterize and
    used gradients to more efficiently increase values while scanning
    the screen.
  - Added an unsafe block in the big rasterize loop.

A lot of comments have been introduced, but I'm pushing them so I can
maybe come back to it later. The next commit should not have them
anymore.

+++ src/graphics/window.rs
@@ -9,9 +9,9 @@ use winit::window;
 /// Contains the necessary information to move and create a window.
 pub struct Window {
     /// Width of the window.
-    width: u32,
+    width: usize,
     /// Height of the window.
-    height: u32,
+    height: usize,
     /// Shared instance of the winit window.
     winit_window: Option<Arc<window::Window>>,
 }
@@ -29,7 +29,7 @@ impl Window {
     /// # Returns
     ///
     /// The instantiated Window.
-    pub fn new(width: u32, height: u32) -> Window {
+    pub fn new(width: usize, height: usize) -> Window {
         Window {
             width: width,
             height: height,
@@ -61,11 +61,11 @@ impl Window {
         Ok(())
     }
     /// Getter for the width.
-    pub fn width(&self) -> u32{
+    pub fn width(&self) -> usize{
         self.width
     }
     /// getter for the height.
-    pub fn height(&self) -> u32{
+    pub fn height(&self) -> usize{
         self.height
     }
     /// Mutable getter for the winit window.
